Fix updates

+++ src/main/java/org/pio/database/EnemyDatabase.java
@@ -4,5 +4,5 @@ public interface EnemyDatabase extends EntityDatabase{
     int getHealth(String fileName);
     int getDamage(String fileName);
     int getGold(String fileName);
-    int getMovementSpeed(String fileName);
+    double getMovementSpeed(String fileName);
 }
