+++ src/main/java/org/pio/entities/enemy/Enemy_2.java
@@ -7,7 +7,7 @@ import java.util.LinkedList;
 
 public class Enemy_2 extends Enemy {
 
-    public Enemy_2(String name, int id, int health, int damage, int gold, int movementSpeed, int width, int height, LinkedHashMap<Directions, LinkedList<String>> sprites) {
+    public Enemy_2(String name, int id, int health, int damage, int gold, double movementSpeed, int width, int height, LinkedHashMap<Directions, LinkedList<String>> sprites) {
         super(name, id, health, damage, gold, movementSpeed, width, height, sprites);
     }
     public Enemy_2(Enemy enemy, int posX, int posY, Directions direction){
