+++ src/main/java/org/pio/entities/enemy/Enemy_3.java
@@ -6,7 +6,7 @@ import java.util.LinkedHashMap;
 import java.util.LinkedList;
 
 public class Enemy_3 extends Enemy{
-    public Enemy_3(String name, int id, int health, int damage, int gold, int movementSpeed, int width, int height, LinkedHashMap<Directions, LinkedList<String>> sprites) {
+    public Enemy_3(String name, int id, int health, int damage, int gold, double movementSpeed, int width, int height, LinkedHashMap<Directions, LinkedList<String>> sprites) {
         super(name, id, health, damage, gold, movementSpeed, width, height, sprites);
     }
     public Enemy_3(Enemy enemy, int posX, int posY, Directions direction){
