+++ src/main/java/org/pio/scene/PlayScene.java
@@ -150,8 +150,8 @@ public class PlayScene extends GameScene implements sceneMeethods, mouseMethods 
             sidePanel.selectedTower.posX=x;
             sidePanel.selectedTower.posY=y;
 
-            sidePanel.selectedTower.bounds.x=sidePanel.selectedTower.posX;
-            sidePanel.selectedTower.bounds.y=sidePanel.selectedTower.posY;
+            sidePanel.selectedTower.bounds.x= (int) sidePanel.selectedTower.posX;
+            sidePanel.selectedTower.bounds.y= (int) sidePanel.selectedTower.posY;
 
             int ellipseOffset=20;
             sidePanel.selectedTower.rangeEllipse.setFrame(sidePanel.selectedTower.posX-sidePanel.selectedTower.range+ellipseOffset, sidePanel.selectedTower.posY-sidePanel.selectedTower.range+ellipseOffset, sidePanel.selectedTower.range*2, sidePanel.selectedTower.range*2);
